requirements.txt
numpy==1.17.3
pandas==0.25.2
pytest==5.2.2
PyYAML==5.1.2
scikit-learn==0.21.3
scipy==1.3.1
tabulate==0.8.5
tensorboard==2.0.0
tensorflow-gpu==2.1.0
termcolor==1.1.0
terminaltables==3.1.0
tqdm==4.36.1
colorama==0.4.3
